Move over code from happo-cypress
This commit moves most of the code in happo-cypress/src into this new
happo-e2e package. I've refactored things a little to make the API a
little cleaner. There are three main modules/scripts that you can use
from happo-e2e:

- happo-e2e -- the CLI formerly known as happo-cypress
- happo-e2e/controller -- the "server" side module, designed to be run
  in a server environment. It's a class that you instantiate.
- happo-e2e/takeDOMSnapshot -- the "client" side module, designed to be
  run in the browser

The idea is that happo-playwright (currently in development) should be
able to make use of this library as well.