Make dynamic svg images appear in screenshots
This commit fixes two things:
- When svg images are part of an asset package, the hash url needs to
  have a file suffix for Happo to show it
- Dynamic asset paths (i.e. containing query params) weren't injected in
  the DOM snapshot html correctly.

The former issue I fixed by mapping the fetched content type to a file
suffix. This is for svg files mainly but I thought it would be to have
all images get the suffix.

The latter issue was fixed by ampersanding "&" before injecting the new
url in the html.