Add support for buffer images
Before, the only way you could upload images was via a path to a file in
the file system. When exploring Appium integrations
(https://github.com/happo/happo-appium) I noticed that we get base64
encoded buffers instead. So to avoid having to save a buffer to file
first, I'm making registerLocalSnapshot and uploadImage work with
buffers as well.

If the width and height is provided and you provide a buffer, we will
now also compute the dimensions on the fly for you.