Add support for HAPPO_BEFORE_SHA_TAG_MATCHER
This environment variable can be used to resolve the previous/before sha
from git tags. If e.g. HAPPO_BEFORE_SHA_TAG_MATCHER=happo-*, the before
sha will be resolved to the commit sha of a previous commit that has a
happo-<something> tag present.

The current commit is ignored when resolving the previous sha. This way
we can retrigger a run for a build that includes tagging the current
commit, without causing it to compare with itself.