Fix adding paths with escape characters to asset package
I came across an example where an image was in the asset package but
wasn't part of the screenshot. It was because the url contained %20 and
that ended up being stored in the folder structure for the image in the
asset package. We can remove the escaped chars so that the image is
later found when referenced in an html document.