Apply css blocks conditionally
Customer Oberst reached out about a heading image looking too big in
their Cypress Happo test suite. I tracked it down to being caused by a
global style for `img` that was part of a previous page. Since
happo-e2e combines all css into a single block, things like this can
happen.

To fix this, I considered two options:
A) Limit one snap-request to a single URL
B) Use conditional styles and apply conditionally depending on page

A would mean having to run `controller.finish()` and `controller.init()`
if we detect the page URL changing. We could ignore query params so that
anything with the same path of the URL would be considered the same
page.

B, which is the option I went with, meant recording all styles, giving
each block an ID and then remember what styles should be part of each
snapshot. Basically the option described here:
https://docs.happo.io/docs/examples#conditionally-applied-stylesheets

B ended up being easier to implement and test, so that's what I went
with. It also means we have fewer assumptions about what the app is
doing. Styles used at a specific time will be associated with a
snapshot, which should be robust.