Allow multiple elements in takeDOMSnapshot.js
This commit addresses an issue where all elements matching a selector
should be included in a screenshot. Specifically, this time it's a
header where certain elements are rendered in a "portal" outside of some
element. By allowing a new option passed in happo-cypress,
"includeAllElements: true", we can send an array of elements to
happo-e2e, and takeDOMSnapshot.js will then combine html for all matches
in a single "package".