Add controller.removeDuplicatesInTimeframe function
This function will be used by happo-cypress to clean up after a retried
spec. In Cypress 12 and earlier, we were able to the the start and end
of the failed attempt but in Cypress 13 we only get the timestamp for
the whole spec execution (including failed attempt + retry). To make the
cleanup process simpler, I'm adding a utility function to remove
duplicates happening in a specific timeframe.