Retry fetch requests
We are seeing some ECONNRESET errors that seem to be happening when we
are fetching CSS assets. Currently this causes the job to fail, but I am
hoping we can smooth over this by adding some retries. Network flake is
inevitable.

This also moves the `res.ok` handling into the fetch function, which
allows us to clean up some other code a bit.

This is similar to how we have makeRequest structured in the happo.io
package:

https://github.com/happo/happo.io/blob/74b23bf4/src/makeRequest.js