Make archives deterministic
This is similar to:

https://github.com/happo/happo.io/pull/292

I believe that we want to set statConcurrency to 1 here so that the
files in the archives end up in the same order every time (assuming the
glob itself is deterministically ordered).

I am updating happo.io at the same time since it brings the same new
version of archiver as well.