Give replaced _external urls a file suffix
I noticed that when a test suite has an ifram with an inlined html file
as a _external/<hash> url, the browser will download the html file as an
attachment instead of using it directly for the iframe. We can fix this
by adding a file suffix to the file when we store it in the assets
package.