Add items to archive serially when creating asset packages
I noticed that our test for this function was flaky. I believe this is
because we were adding items to the archive in our Promise.all, which
could cause items to be added in an inconsistent order. We really want
this to be built deterministically, so I am splitting it out so that we
fetch the items in parallel and then follow up with adding them to the
archive serially. This should probuce a more stable output.