Inline shadow roots so that workers can reproduce them
This is an attempt at fixing one of the most disabling issues with our
Cypress and Playwright integrations. Since we rely on DOM snapshots,
supporting custom elements has turned out to be tricky. To fix that, I'm
doing the following:

- Detect shadow roots as part of the DOM snapshot
- Inline the shadow root html as a hidden element (happo-shadow-content)

On the Happo worker, we will
- Query the DOM for potential `happo-shadow-content` elements
- Create a shadow root and inject the shadow content in the new shadow
  root.

It's very possible that I've missed some details of how shadow roots and
custom elements work. We can address them as they show up.