Add support for <link> and adopted stylesheets
So far we've only handled <style> elements inside shadow roots. That's
not the only way you can style shadow roots however. We have to handle
<link href> and adopted stylesheets as well.

For <link href>, add the asset to the list of assetUrls. This will make
it part of the asset package and in turn make it available on the Happo
workers.

For adopted stylesheets, inline them as <style> elements.