Improve error message when unable to resolve GITHUB_EVENT_PATH
The error message here was a bit cryptic:

```
Error: Cannot find module '/path/to/some/file.json'
Require stack:
- /e2e/node_modules/happo-e2e/src/resolveEnvironment.js
...
```

I want to make it easier for people to troubleshoot this on their own. I
think we can try/catch the require and provide a clearer error message
that helps people understand that something may be up with the
GITHUB_EVENT_PATH environment variable.