Always include svgs with <symbol> elements
When a page uses `<use>` elements referencing svg snippets defined
elsewere on the page, we weren't including the source `<symbol>`s. By
always injecting html for all svgs that define the `<symbol>` element, I
think we can get these svgs to appear correctly in the screenshots, even
when the svgs are defined outside of the element currently beeing
screenshot.