Add support for `strategy: 'clip'` for takeDOMSnapshot
This option will change the behavior of the DOM snapshot. Instead of
only including the element in the resulting html `strategy: 'clip'` will
include the entire body, but will leave a data attribute in the DOM for
the element under test.

Happo workers will honor this attribute and clip the screenshot on that
element.