Use textContent of stylesheet if available
This will fix a bug where our style collection would not work if the
stylesheet was using this pattern:

  font: var(--some-var);
  font-weight: var(--some-other-var);

Iterating over cssRules here would mean we collected styles that looked
something like:

  font-style: ;
  font-variant-ligatures: ;
  font-variant-caps: ;
  font-variant-numeric: ;
  font-variant-east-asian: ;
  font-variant-alternates: ;
  font-variant-position: ;
  font-variant-emoji: ;
  font-stretch: ;
  font-size: ;
  line-height: ;
  font-family: ;
  font-optical-sizing: ;
  font-size-adjust: ;
  font-kerning: ;
  font-feature-settings: ;
  font-variation-settings: ;
  font-weight: var(--some-other-var);

Notice how properties are expanded and left with no content.

To fix this, we can use `textContent` of the style element instead. And
if that isn't available, we fall back to `sheet.cssRules`.

The only potential issue this could introduce is if you mix text content
styles and dynamically injected ones (using e.g. `insertRule`). I doubt
that's a common way to create styling however.