Localize test setup
I wanted to run the controller test on its own, but it failed because
these env vars were set in the shell script. I started by adding a
setup.js file and importing it in all of the tests but then I realized
that we only need to do this in the controller test so I moved it there.
To make this work, we need to read the env vars lazily and not in the
module scope, so I did some light refactoring as well.