Handle broken CSS files when there are external assets
When there are external assets at the same time as CSS files that fail
to download, the CSS file will have an undefined `.css` property. We did
not guard against that here, so people in this scenario ended up running
into an unhandled error.

To fix this, we just need to guard against block.css being undefined.

I was able to reproduce this in our test suite and verified that it
fixes the bug.