Enable HAPPO_GITHUB_USER_CREDENTIALS to post comments
We do this for other Happo integrations. Documentation:

  https://docs.happo.io/docs/continuous-integration#posting-statuses-without-installing-the-happo-github-app

The relevant code is here:

  https://github.com/happo/happo.io/blob/5c257d54/src/executeCli.js#L133-L144

I noticed in that code that we do not post this comment if Happo async
is enabled. I also know that happo-e2e only uses Happo async. It looks
like this was done when async reports were initially added here:

  https://github.com/happo/happo.io/pull/155/files#diff-fef5bcd034e4f4b70bb46ab914a3a4d084c58930769d1b795c152aa2dbf7f0dcR117

It isn't clear if there was a good reason for this, or if it was an
oversight. For now I am proceeding with the hope that it will work just
fine for async reports and comparisons.