Add first test for bin/happo-e2e.js
We recently shipped a bug with an incorrect import statement. This
should have been caught by any test running on this file but we did not
have any.

I am getting the ball rolling here by writing a very basic test on this
file. To make this bin script importable and executable as a function,
I am doing some light refactoring at the same time.

We should definitely come back and add more testing, but for now this is
a great improvement.