Add monkey-patch for constructed styles
We know that using shorthand values with variables can lead to issues
with the DOM snapshot. For instance,

p {
  font: var(--font);
  font-weight: var(--font-weight);
}

becomes

p {
  font-family: ;
  font-weight: ;
  font-size: ;
  ... etc
}

See:
https://issues.chromium.org/issues/40905071#comment7
https://issues.chromium.org/issues/40240778#comment5

We can fix this by monkey-patching CSSStyleSheet and its functions. By
recording the strings exactly as they are injected, we can preserve the
"real" css that the user intended to apply.