Allow applyConstructedStylesPatch to take a window argument
We want this patch to work in Cypress, where the page is rendered in an
iframe and we need to run our patch on a different `window`. To make
this work, happo-e2e needs to expose the patch function in a way that it
can be applied to a different `window`, so we are adding a window
argument to the function. It defaults to `window`, so it will continue
working for the current use cases just fine.

Cypress needs to call this function in its own special Cypress way, so
we are adding a property to takeDOMSnapshot to make importing it easier
and backwards compatible.