Override previous snapshot when deduping
The previous behavior on encountering a duplicate (same component and
variant) snapshot was to append a "-2" suffix (or -3, -4 and so on).
This was originally added in this commit as a way to allow people to
call happoScreenshot() without any arguments:
https://github.com/happo/happo-cypress/commit/8eae597414bbd9f6212362a672dd5653c35de263

I believe this behavior isn't the right one. Instead we can simply
override the previously registered snapshot with the same identifier.
This will make cases where a test case is retried better, as we won't
confusingly create two snapshots just because the test was (auto)
retried.

This will be a breaking change.